requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.5
bcrypt==5.0.0
bidict==0.23.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cryptography==46.0.3
datasets==4.4.1
dill==0.4.0
docstring_parser==0.17.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
google-adk==1.18.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.47.0
google-generativeai==0.8.2
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
multiprocess==0.70.18
numpy==2.0.2
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycountry==24.6.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.14.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
starlette==0.49.3
tenacity==9.1.2
tomli==2.3.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
PyJWT
google-auth-oauthlib
email-validator
itsdangerous>=2.2.0
mypy==1.11.1
